+++ src/components/DisplayFull.tsx
