+++ src/components/DisplayPreview.tsx
