+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { BrowserRouter as Router } from 'react-router-dom';
-import App from './App.jsx'
+import App from './App.js'
 
 import './index.css'
