+++ src/utils/isTimePeriodValid.ts
